+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Here is the list of the changes and features provided by the **Rainbow-Node-SDK*
 All notable changes to Rainbow-Node-SDK will be documented in this file.
 
 ## [1.59.0] - 2018-08-xx
--   Fix : 
+-   Add in AdminService the method to get all users for a given admin
 
 ## [1.58.0] - 2018-07-10
 -   Fix : event presence rainbow_oncontactpresencechanged when a contact in the roster is on phone.